Update prime logic.

+++ Topics/Java-8-Features/Lambda-Expressions/Lambda1.java
@@ -8,7 +8,7 @@ public class Lambda1 {
 
         arrL.forEach(num -> {
             boolean isPrime = true;
-            for (int i = 2; i <= (int) Math.sqrt(num); i++) {
+            for (int i = 2; i <= Math.sqrt(num); i++) {
                 if (num % i == 0) {
                     isPrime = false;
                     break;
